refactor: SRP를 고려해 Quest API 분리
