+++ src/main/java/com/groom/orbit/goal/quest/controller/QuestUpdateController.java
@@ -1,0 +1,27 @@
+package com.groom.orbit.goal.quest.controller;
+
+import org.springframework.web.bind.annotation.*;
+
+import com.groom.orbit.common.annotation.AuthMember;
+import com.groom.orbit.common.dto.CommonSuccessDto;
+import com.groom.orbit.common.dto.ResponseDto;
+import com.groom.orbit.goal.quest.application.QuestUpdateService;
+import com.groom.orbit.goal.quest.application.dto.request.UpdateQuestRequestDto;
+
+import lombok.RequiredArgsConstructor;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/quest")
+public class QuestUpdateController {
+
+  private final QuestUpdateService questUpdateService;
+
+  @PatchMapping("/{quest_id}")
+  public ResponseDto<CommonSuccessDto> updateQuest(
+      @AuthMember Long memberId,
+      @PathVariable("quest_id") Long questId,
+      @RequestBody UpdateQuestRequestDto dto) {
+    return ResponseDto.ok(questUpdateService.updateQuest(memberId, questId, dto));
+  }
+}
